Update SDKs (iOS to 9, Android to 20)
The addition of the simulator identifier to the test ads devices list
is removed because the code was broken, and the documentation states
that simulators are treated as test devices automatically.

Replaced the previous consent libraries with UserMessagingPlatform for
both platforms. The privacy url is specified in the online dashboard
now, so that block is removed.